migrate GetInspection (RAW)
