--- pkg/internal/inspections/get_inspection.go
@@ -1,40 +1,0 @@
-package inspections
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-
-	"github.com/SafetyCulture/safetyculture-exporter/pkg/httpapi"
-	"github.com/SafetyCulture/safetyculture-exporter/pkg/internal/util"
-	"github.com/pkg/errors"
-)
-
-// GetInspection retrieves the inspection of the given id.
-func GetInspection(ctx context.Context, apiClient *httpapi.Client, id string) (*json.RawMessage, error) {
-	var (
-		result *json.RawMessage
-		errMsg json.RawMessage
-	)
-
-	sl := apiClient.Sling.New().Get(fmt.Sprintf("/audits/%s", id)).
-		Set(string(httpapi.Authorization), apiClient.AuthorizationHeader).
-		Set(string(httpapi.IntegrationID), apiClient.IntegrationID).
-		Set(string(httpapi.IntegrationVersion), apiClient.IntegrationVersion).
-		Set(string(httpapi.XRequestID), util.RequestIDFromContext(ctx))
-
-	req, _ := sl.Request()
-	req = req.WithContext(ctx)
-
-	_, err := apiClient.Do(&util.SlingHTTPDoer{
-		Sl:       sl,
-		Req:      req,
-		SuccessV: &result,
-		FailureV: &errMsg,
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "api request")
-	}
-
-	return result, nil
-}
